Set resizable to false again
It already is resizable by using the scale factor, and it messes
up my i3wm system because then it automatically changes the
size to fit with the other windows. It also doesn't actually resize.
I assume this is the reason it was put there in the first place.

I'm guessing this was put here to scale to a full
monitor, which should be further looked at.